rename previous bot to bot_impl
