+++ bot_impl.py
